Cleaned up UI code a little more
